backend/requirements.txt
fastapi==0.121.3
uvicorn[standard]==0.38.0
python-dotenv==1.2.1
pydantic==2.12.4
pydantic-settings==2.7.1
email-validator==2.1.0

# Graph computation
networkx==3.5
igraph==0.11.8
numpy==2.2.0
scipy==1.15.0

# Data processing
polars==1.35.2

# Database & Cache
redis[hiredis]==7.1.0
motor==3.7.0
beanie==1.27.0
PyJWT[crypto]==2.9.0
argon2-cffi==23.1.0
python-multipart==0.0.20
aiofiles==24.1.0
orjson==3.10.12
loguru==0.7.3
